+++ PlayHouse/PlayHouse/Service/Session/network/PacketParser.cs
@@ -41,7 +41,7 @@ namespace PlayHouse.Service.Session.network
                     buffer.Clear(2);
 
                     short serviceId = XBitConverter.ToHostOrder(buffer.ReadInt16());
-                    short msgId = XBitConverter.ToHostOrder(buffer.ReadInt16());
+                    int msgId = XBitConverter.ToHostOrder(buffer.ReadInt32());
                     short msgSeq = XBitConverter.ToHostOrder(buffer.ReadInt16());
                     byte stageId = buffer.ReadByte();
